+++ packages/navigation/atlassian-switcher/src/primitives/index.tsx
@@ -1,0 +1,7 @@
+import SwitcherItem from './item';
+import Section from './section';
+import SwitcherWrapper from './wrapper';
+import ManageButton from './manage-button';
+import Skeleton from './skeleton';
+
+export { SwitcherItem, SwitcherWrapper, ManageButton, Section, Skeleton };