+++ packages/navigation/atlassian-switcher/CHANGELOG.md
@@ -1,6 +1,6 @@
-# centralised-app-switcher
+# @atlaskit/atlassian-switcher
 
 ## 1.0.1
 - [patch] [84228f6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/84228f6):
 
-  - Centralised app switcher initial commit
+  - Atlassian switcher initial commit